fix (BugSplatManager): unregister logMessageReceived on destruction (#94)
Application.logMessageReceived is a static callback handler, which isn't reset when changing scenes or when exiting play mode if domain reloads are not enabled.

When destroying the class without unregistering, e.g. due to scene change or existing play mode, the callback will then try to start a coroutine on the destroyed object.